final/papaya2/requirements.txt
APScheduler==3.9.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
cryptography==36.0.1
DateTime==4.3
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-ReCaptcha==0.4.2
Flask-Session==0.4.0
Flask-WTF==1.0.1
itsdangerous==2.0.1
Jinja2==3.0.2
MyCaesarCipher==0.1.0
Pillow==9.0.0
pycryptodome==3.14.1
pyotp==2.6.0
requests==2.28.1
shortuuid==1.0.9
WTForms==3.0.1
secure
nvdlib
